[Feature] Skip compiling the non-supported cmdlets in S.M.A.dll
This is to keep the cache in InitialSessionState in sycn with the actual available cmdlet types in System.Management.Automation.dll
