+++ src/System.Management.Automation/engine/remoting/commands/ConnectPSSession.cs
@@ -9,7 +9,6 @@ using System.Management.Automation;
 using System.Management.Automation.Internal;
 using System.Management.Automation.Remoting;
 using System.Management.Automation.Runspaces;
-using System.Management.Automation.Remoting.Client;
 using System.Management.Automation.Host;
 using System.Threading;
 using Dbg = System.Management.Automation.Diagnostics;
@@ -1087,425 +1086,4 @@ namespace Microsoft.PowerShell.Commands
 
         #endregion
     }
-
-    #region QueryRunspaces
-
-    internal class QueryRunspaces
-    {
-        #region Constructor
-
-        internal QueryRunspaces()
-        {
-            _stopProcessing = false;
-        }
-
-        #endregion
-
-        #region Internal Methods
-
-        /// <summary>
-        /// Queries all remote computers specified in collection of WSManConnectionInfo objects
-        /// and returns disconnected PSSession objects ready for connection to server.
-        /// Returned sessions can be matched to Guids or Names.
-        /// </summary>
-        /// <param name="connectionInfos">Collection of WSManConnectionInfo objects.</param>
-        /// <param name="host">Host for PSSession objects.</param>
-        /// <param name="stream">Out stream object.</param>
-        /// <param name="runspaceRepository">Runspace repository.</param>
-        /// <param name="throttleLimit">Throttle limit.</param>
-        /// <param name="filterState">Runspace state filter value.</param>
-        /// <param name="matchIds">Array of session Guids to match to.</param>
-        /// <param name="matchNames">Array of session Names to match to.</param>
-        /// <param name="configurationName">Configuration name to match to.</param>
-        /// <returns>Collection of disconnected PSSession objects.</returns>
-        internal Collection<PSSession> GetDisconnectedSessions(Collection<WSManConnectionInfo> connectionInfos, PSHost host,
-                                                               ObjectStream stream, RunspaceRepository runspaceRepository,
-                                                               int throttleLimit, SessionFilterState filterState,
-                                                               Guid[] matchIds, string[] matchNames, string configurationName)
-        {
-            Collection<PSSession> filteredPSSessions = new Collection<PSSession>();
-
-            // Create a query operation for each connection information object.
-            foreach (WSManConnectionInfo connectionInfo in connectionInfos)
-            {
-                Runspace[] runspaces = null;
-
-                try
-                {
-                    runspaces = Runspace.GetRunspaces(connectionInfo, host, BuiltInTypesTable);
-                }
-                catch (System.Management.Automation.RuntimeException e)
-                {
-                    if (e.InnerException is InvalidOperationException)
-                    {
-                        // The Get-WSManInstance cmdlet used to query remote computers for runspaces will throw
-                        // an Invalid Operation (inner) exception if the connectInfo object is invalid, including
-                        // invalid computer names.
-                        // We don't want to propagate the exception so just write error here.
-                        if (stream.ObjectWriter != null && stream.ObjectWriter.IsOpen)
-                        {
-                            int errorCode;
-                            string msg = StringUtil.Format(RemotingErrorIdStrings.QueryForRunspacesFailed, connectionInfo.ComputerName, ExtractMessage(e.InnerException, out errorCode));
-                            string FQEID = WSManTransportManagerUtils.GetFQEIDFromTransportError(errorCode, "RemotePSSessionQueryFailed");
-                            Exception reason = new RuntimeException(msg, e.InnerException);
-                            ErrorRecord errorRecord = new ErrorRecord(reason, FQEID, ErrorCategory.InvalidOperation, connectionInfo);
-                            stream.ObjectWriter.Write((Action<Cmdlet>)(cmdlet => cmdlet.WriteError(errorRecord)));
-                        }
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-
-                if (_stopProcessing)
-                {
-                    break;
-                }
-
-                // Add all runspaces meeting filter criteria to collection.
-                if (runspaces != null)
-                {
-                    // Convert configuration name into shell Uri for comparison.
-                    string shellUri = null;
-                    if (!string.IsNullOrEmpty(configurationName))
-                    {
-                        shellUri = (configurationName.IndexOf(
-                                    System.Management.Automation.Remoting.Client.WSManNativeApi.ResourceURIPrefix, StringComparison.OrdinalIgnoreCase) != -1) ?
-                                    configurationName : System.Management.Automation.Remoting.Client.WSManNativeApi.ResourceURIPrefix + configurationName;
-                    }
-
-                    foreach (Runspace runspace in runspaces)
-                    {
-                        // Filter returned runspaces by ConfigurationName if provided.
-                        if (shellUri != null)
-                        {
-                            // Compare with returned shell Uri in connection info.
-                            WSManConnectionInfo wsmanConnectionInfo = runspace.ConnectionInfo as WSManConnectionInfo;
-                            if (wsmanConnectionInfo != null &&
-                                !shellUri.Equals(wsmanConnectionInfo.ShellUri, StringComparison.OrdinalIgnoreCase))
-                            {
-                                continue;
-                            }
-                        }
-
-                        // Check the repository for an existing viable PSSession for
-                        // this runspace (based on instanceId).  Use the existing
-                        // local runspace instead of the one returned from the server
-                        // query.
-                        PSSession existingPSSession = null;
-                        if (runspaceRepository != null)
-                        {
-                            existingPSSession = runspaceRepository.GetItem(runspace.InstanceId);
-                        }
-
-                        if (existingPSSession != null &&
-                            UseExistingRunspace(existingPSSession.Runspace, runspace))
-                        {
-                            if (TestRunspaceState(existingPSSession.Runspace, filterState))
-                            {
-                                filteredPSSessions.Add(existingPSSession);
-                            }
-                        }
-                        else if (TestRunspaceState(runspace, filterState))
-                        {
-                            filteredPSSessions.Add(new PSSession(runspace as RemoteRunspace));
-                        }
-                    }
-                }
-            }
-
-            // Return only PSSessions that match provided Ids or Names.
-            if ((matchIds != null) && (filteredPSSessions.Count > 0))
-            {
-                Collection<PSSession> matchIdsSessions = new Collection<PSSession>();
-                foreach (Guid id in matchIds)
-                {
-                    bool matchFound = false;
-                    foreach (PSSession psSession in filteredPSSessions)
-                    {
-                        if (_stopProcessing)
-                        {
-                            break;
-                        }
-
-                        if (psSession.Runspace.InstanceId.Equals(id))
-                        {
-                            matchFound = true;
-                            matchIdsSessions.Add(psSession);
-                            break;
-                        }
-                    }
-
-                    if (!matchFound && stream.ObjectWriter != null && stream.ObjectWriter.IsOpen)
-                    {
-                        string msg = StringUtil.Format(RemotingErrorIdStrings.SessionIdMatchFailed, id);
-                        Exception reason = new RuntimeException(msg);
-                        ErrorRecord errorRecord = new ErrorRecord(reason, "PSSessionIdMatchFail", ErrorCategory.InvalidOperation, id);
-                        stream.ObjectWriter.Write((Action<Cmdlet>)(cmdlet => cmdlet.WriteError(errorRecord)));
-                    }
-                }
-
-                // Return all found sessions.
-                return matchIdsSessions;
-            }
-            else if ((matchNames != null) && (filteredPSSessions.Count > 0))
-            {
-                Collection<PSSession> matchNamesSessions = new Collection<PSSession>();
-                foreach (string name in matchNames)
-                {
-                    WildcardPattern namePattern = WildcardPattern.Get(name, WildcardOptions.IgnoreCase);
-                    bool matchFound = false;
-                    foreach (PSSession psSession in filteredPSSessions)
-                    {
-                        if (_stopProcessing)
-                        {
-                            break;
-                        }
-
-                        if (namePattern.IsMatch(((RemoteRunspace)psSession.Runspace).RunspacePool.RemoteRunspacePoolInternal.Name))
-                        {
-                            matchFound = true;
-                            matchNamesSessions.Add(psSession);
-                        }
-                    }
-
-                    if (!matchFound && stream.ObjectWriter != null && stream.ObjectWriter.IsOpen)
-                    {
-                        string msg = StringUtil.Format(RemotingErrorIdStrings.SessionNameMatchFailed, name);
-                        Exception reason = new RuntimeException(msg);
-                        ErrorRecord errorRecord = new ErrorRecord(reason, "PSSessionNameMatchFail", ErrorCategory.InvalidOperation, name);
-                        stream.ObjectWriter.Write((Action<Cmdlet>)(cmdlet => cmdlet.WriteError(errorRecord)));
-                    }
-                }
-
-                return matchNamesSessions;
-            }
-            else
-            {
-                // Return all collected sessions.
-                return filteredPSSessions;
-            }
-        }
-
-        /// <summary>
-        /// Returns true if the existing runspace should be returned to the user
-        /// a.  If the existing runspace is not broken
-        /// b.  If the queried runspace is not connected to a different user.
-        /// </summary>
-        /// <param name="existingRunspace"></param>
-        /// <param name="queriedrunspace"></param>
-        /// <returns></returns>
-        private static bool UseExistingRunspace(
-            Runspace existingRunspace,
-            Runspace queriedrunspace)
-        {
-            Dbg.Assert(existingRunspace != null, "Invalid parameter.");
-            Dbg.Assert(queriedrunspace != null, "Invalid parameter.");
-
-            if (existingRunspace.RunspaceStateInfo.State == RunspaceState.Broken)
-            {
-                return false;
-            }
-
-            if (existingRunspace.RunspaceStateInfo.State == RunspaceState.Disconnected &&
-                queriedrunspace.RunspaceAvailability == RunspaceAvailability.Busy)
-            {
-                return false;
-            }
-
-            // Update existing runspace to have latest DisconnectedOn/ExpiresOn data.
-            existingRunspace.DisconnectedOn = queriedrunspace.DisconnectedOn;
-            existingRunspace.ExpiresOn = queriedrunspace.ExpiresOn;
-
-            return true;
-        }
-
-        /// <summary>
-        /// Returns Exception message.  If message is WSMan Xml then
-        /// the WSMan message and error code is extracted and returned.
-        /// </summary>
-        /// <param name="e">Exception</param>
-        /// <param name="errorCode">Returned WSMan error code</param>
-        /// <returns>WSMan message</returns>
-        internal static string ExtractMessage(
-            Exception e,
-            out int errorCode)
-        {
-            errorCode = 0;
-
-            if (e == null ||
-                e.Message == null)
-            {
-                return string.Empty;
-            }
-
-            string rtnMsg = null;
-            try
-            {
-                System.Xml.XmlReaderSettings xmlReaderSettings = InternalDeserializer.XmlReaderSettingsForUntrustedXmlDocument.Clone();
-                xmlReaderSettings.MaxCharactersInDocument = 4096;
-                xmlReaderSettings.MaxCharactersFromEntities = 1024;
-                xmlReaderSettings.DtdProcessing = System.Xml.DtdProcessing.Prohibit;
-
-                using (System.Xml.XmlReader reader = System.Xml.XmlReader.Create(
-                        new System.IO.StringReader(e.Message), xmlReaderSettings))
-                {
-                    while (reader.Read())
-                    {
-                        if (reader.NodeType == System.Xml.XmlNodeType.Element)
-                        {
-                            if (reader.LocalName.Equals("Message", StringComparison.OrdinalIgnoreCase))
-                            {
-                                rtnMsg = reader.ReadElementContentAsString();
-                            }
-                            else if (reader.LocalName.Equals("WSManFault", StringComparison.OrdinalIgnoreCase))
-                            {
-                                string errorCodeString = reader.GetAttribute("Code");
-                                if (errorCodeString != null)
-                                {
-                                    try
-                                    {
-                                        // WinRM returns both signed and unsigned 32 bit string values.  Convert to signed 32 bit integer.
-                                        Int64 eCode = Convert.ToInt64(errorCodeString, System.Globalization.NumberFormatInfo.InvariantInfo);
-                                        unchecked
-                                        {
-                                            errorCode = (int)eCode;
-                                        }
-                                    }
-                                    catch (FormatException)
-                                    { }
-                                    catch (OverflowException)
-                                    { }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            catch (System.Xml.XmlException)
-            { }
-
-            return rtnMsg ?? e.Message;
-        }
-
-        /// <summary>
-        /// Discontinue all remote server query operations.
-        /// </summary>
-        internal void StopAllOperations()
-        {
-            _stopProcessing = true;
-        }
-
-        /// <summary>
-        /// Compares the runspace filter state with the runspace state.
-        /// </summary>
-        /// <param name="runspace">Runspace object to test.</param>
-        /// <param name="filterState">Filter state to compare.</param>
-        /// <returns>Result of test.</returns>
-        public static bool TestRunspaceState(Runspace runspace, SessionFilterState filterState)
-        {
-            bool result;
-
-            switch (filterState)
-            {
-                case SessionFilterState.All:
-                    result = true;
-                    break;
-
-                case SessionFilterState.Opened:
-                    result = (runspace.RunspaceStateInfo.State == RunspaceState.Opened);
-                    break;
-
-                case SessionFilterState.Closed:
-                    result = (runspace.RunspaceStateInfo.State == RunspaceState.Closed);
-                    break;
-
-                case SessionFilterState.Disconnected:
-                    result = (runspace.RunspaceStateInfo.State == RunspaceState.Disconnected);
-                    break;
-
-                case SessionFilterState.Broken:
-                    result = (runspace.RunspaceStateInfo.State == RunspaceState.Broken);
-                    break;
-
-                default:
-                    Dbg.Assert(false, "Invalid SessionFilterState value.");
-                    result = false;
-                    break;
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Returns the default type table for built-in PowerShell types.
-        /// </summary>
-        internal static TypeTable BuiltInTypesTable
-        {
-            get
-            {
-                if (s_TypeTable == null)
-                {
-                    lock (s_SyncObject)
-                    {
-                        if (s_TypeTable == null)
-                        {
-                            s_TypeTable = TypeTable.LoadDefaultTypeFiles();
-                        }
-                    }
-                }
-
-                return s_TypeTable;
-            }
-        }
-
-        #endregion
-
-        #region Private Members
-
-        private bool _stopProcessing;
-
-        private static readonly object s_SyncObject = new object();
-        private static TypeTable s_TypeTable;
-
-        #endregion
-    }
-
-    #endregion
-
-    # region Public SessionFilterState Enum
-
-    /// <summary>
-    /// Runspace states that can be used as filters for querying remote runspaces.
-    /// </summary>
-    public enum SessionFilterState
-    {
-        /// <summary>
-        /// Return runspaces in any state.
-        /// </summary>
-        All = 0,
-
-        /// <summary>
-        /// Return runspaces in Opened state.
-        /// </summary>
-        Opened = 1,
-
-        /// <summary>
-        /// Return runspaces in Disconnected state.
-        /// </summary>
-        Disconnected = 2,
-
-        /// <summary>
-        /// Return runspaces in Closed state.
-        /// </summary>
-        Closed = 3,
-
-        /// <summary>
-        /// Return runspaces in Broken state.
-        /// </summary>
-        Broken = 4
-    }
-
-    #endregion
 }
